Add Coders/Decoders (IntCoder) and tests
Add some more tests for bugs found (BitSink/BitSource)

+++ Modeller.cpp
@@ -29,7 +29,7 @@ Model getMagBitsModel(int val)
       nbits++;
       val >>= 1;
     }
-    return Model{(uint8_t)nbits, nbits, temp};
+    return Model{nbits, nbits, temp};
 }
 
 class MagModeller : public Modeller
